Move complete tests to seperate package
